% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsDataSource.R
\name{hsDataSource}
\alias{hsDataSource}
\title{Get Metadata About Data Sources}
\usage{
hsDataSource(
  qua.level = NULL,
  moniPoint = NULL,
  parName,
  kind = "q",
  owner = "KWB",
  dbg = FALSE
)
}
\arguments{
\item{qua.level}{data quality level ("r" = raw, "v" = validated, "c" =
calibrated)}

\item{moniPoint}{name of monitoring point, e.g. "STA", "TEG", "MUE"}

\item{parName}{name of parameter, e.g. "AFS", "CSB", "CSBf"}

\item{kind}{kind of data: "q" = water quality, "h" = hydraulic data, "r" =
rain}

\item{owner}{owner of the data, one of "KWB", "SEN", "BWB"}

\item{dbg}{whether to show debug messages or not}
}
\value{
Returns a list with the following named elements: \enumerate{
  \item \code{mdb}: full path to Access database, 
  \item \code{tbl}: table name,
  \item \code{tsField}: name of timestamp field, 
  \item \code{parField}: name of parameter field
}
}
\description{
Returns a list containing \enumerate{
  \item full path to Access database,
  \item table name,
  \item name of timestamp field, 
  \item name of parameter field
} 
for the table that contains data of parameter \code{parName}, measured at
monitoring point \code{moniPoint} in data quality level \code{qua.level}.
}
\examples{
# Get source information of validated data of parameter AFS from Stallstr. 
si <- hsDataSource("v", "STA", "AFS")
si 
# ouput:
#  $db
#  [1] "//moby/miacso$/Daten/ACCESS/KwbMonitoring/2VAL/KWB_STA_VAL.mdb"
# 
#  $tbl
#  [1] "KWB_STA_ScanPar_AFS_VAL"
# 
#  $tsfield
#  [1] "myDateTime"
# 
#  $parfield
#  [1] "AFS_A"

# Access the pieces of information with the $ operator:
si$mdb # [1] "//moby/miacso$/Daten/ACCESS/KwbMonitoring/2VAL/KWB_STA_VAL.mdb"
si$tbl # [1] "KWB_STA_ScanPar_AFS_VAL"
}
