% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miacsoData.R
\name{hsFpFields}
\alias{hsFpFields}
\title{Names of Fields in Fingerprint Table}
\usage{
hsFpFields(moniPoint)
}
\arguments{
\item{moniPoint}{acronym of monitoring point: "STA", "TEG" or "MUE"}
}
\value{
String representing comma separated list of fields in fingerprint
  table
}
\description{
String representing comma separated list of fields in fingerprint table
}
