% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miacsoData.R
\name{hsGetFpAndValRaw}
\alias{hsGetFpAndValRaw}
\title{Read Values and Fingerprints from Database}
\usage{
hsGetFpAndValRaw(moniPoint, parNames, year, firstDate, lastDate,
  home = FALSE)
}
\arguments{
\item{moniPoint}{name of monitoring point}

\item{parNames}{vector of parameter names}

\item{year}{year of which data is requested}

\item{firstDate}{first date of requested time period}

\item{lastDate}{last date of requested time period}

\item{home}{if \code{TRUE}, Hauke's home path is used instead of kwb office
path}
}
\value{
Return the query result as a new data.frame (which is a special type
  of a list) with elements "dts" (date-times), "pars" (parameter values) and
  "fps" (fingerprints)
}
\description{
For the given parameters, the values and fingerprints are read from the
database.
}
