% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miacsoData.R
\name{hsGetValData}
\alias{hsGetValData}
\title{Get Validated Data from Validated Database}
\usage{
hsGetValData(moniPoint, parName, firstDate, lastDate)
}
\arguments{
\item{moniPoint}{name of monitoring point, e.g. "TEG", "STA", "MUE"}

\item{parName}{parameter name, e.g. "AFS"}

\item{firstDate}{first date to be selected as "mm/dd/yyyy hh:nn:ss"}

\item{lastDate}{last date to be selected as "mm/dd/yyyy hh:nn:ss"}
}
\value{
data.frame containing validated data
}
\description{
2011-10-25: created
}
