% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelength.R
\name{hsIndexOfWavelength}
\alias{hsIndexOfWavelength}
\title{Index of Wavelength}
\usage{
hsIndexOfWavelength(wavelength)
}
\arguments{
\item{wavelength}{Wavelength for which corresponding index shall be
determined}
}
\value{
Index corresponding the given wavelength
}
\description{
Returns the index at which the wavelength given in nm can be found in a
vector (i = 1: 200nm, i = 2: 202.5nm, ..., i = 217: 740nm).
}
\details{
2011-12-19: moved from hsLibFingerprint.r
}
