% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelength.R
\name{hsLastWL}
\alias{hsLastWL}
\title{Last Available Wavelength for Given Monitoring Point}
\usage{
hsLastWL(moniPoint)
}
\arguments{
\item{moniPoint}{Name of monitoring point, e.g. "TEG", "STA", "MUE"}
}
\value{
Last wavelength being provided by spectrometer at given monitoring
  point
}
\description{
Returns the last available wavelength of the spectrometer installed at the 
given monitoring point.
}
\details{
2011-12-19: moved from hsLibFingerprint.r
}
