% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsMiaCsoDataAvailability.R
\name{hsMiaCsoDataAvailability}
\alias{hsMiaCsoDataAvailability}
\title{Provide Information on Data Availabilities}
\usage{
hsMiaCsoDataAvailability(
  level,
  moniPoint,
  parName,
  dateFirst = NULL,
  dateLast = NULL,
  tstep = NULL,
  dbg = FALSE
)
}
\arguments{
\item{level}{one of "r" (= raw), "v" (= validated), "c" (= calibrated)}

\item{moniPoint}{one of "STA" (= Stallstr.), "TEG" (= Tegeler Weg), "MUE" (=
Muehlendamm)}

\item{parName}{e.g. "AFS", "CSB", "CSBf", ...}

\item{dateFirst}{Date object representing first date to be considered}

\item{dateLast}{Date object representing last date to be considered}

\item{tstep}{expected time step between time stamps in seconds. Default:
minimum time difference found between consecutive timestamps in given
interval}

\item{dbg}{If \code{TRUE}, debug messages will be shown}
}
\description{
Data availability of raw/validated/calibrated data for parameter \code{para},
measured at monitoring point \code{moniPoint} between \code{dateFirst} and
\code{dateLast}.
}
\examples{
\dontrun{
# Get data availability of raw data of parameter "CSBf", measured at
# monitoring point "TEG" (Tegeler Weg) between 2011-04-01 and 2011-10-01
da <- hsMiaCsoDataAvailability("r", "TEG", "CSBf", "2011-04-01", "2011-10-01")
head(da)
}
# Output:
#   myInterval myCount   myAvail
# 1 2011-04-01    1440 100.00000
# 2 2011-04-02    1440 100.00000
# 3 2011-04-03    1440 100.00000
# 4 2011-04-04    1440 100.00000
# 5 2011-04-05    1440 100.00000
# 6 2011-04-06    1409  97.84722
}
