% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsMoniPoints.R
\name{hsMoniPoints}
\alias{hsMoniPoints}
\title{Names of Available Monitoring Points}
\usage{
hsMoniPoints(kind = NULL, owner = "KWB")
}
\arguments{
\item{kind}{kind of data: "q" = water quality, "h" = hydraulic data, "r" =
rain}

\item{owner}{owner of the data, one of "KWB", "SEN", "BWB"}
}
\description{
Returns the names of available monitoring points
}
