% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsPars.R
\name{hsPars}
\alias{hsPars}
\title{Names of available parameters}
\usage{
hsPars(kind = NULL, moniPoint = NULL, qua.level = "c",
  owner = "KWB", dbg = TRUE)
}
\arguments{
\item{kind}{kind of data: "q" = water quality, "h" = hydraulic data, "r" =
rain}

\item{moniPoint}{name of monitoring point, e.g. "STA", "TEG", "MUE"}

\item{qua.level}{data quality level ("r" = raw, "v" = validated, "c" =
calibrated)}

\item{owner}{owner of the data, one of "KWB", "SEN", "BWB"}

\item{dbg}{whether to show debug messages or not}
}
\value{
Returns the names of available parameters
}
\description{
Names of available parameters
}
