% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsPlotAllDataAvailabilities.R
\name{hsPlotMiaCsoAvailabilities}
\alias{hsPlotMiaCsoAvailabilities}
\title{Plots availability of raw and validated data as bar plot into one plot}
\usage{
hsPlotMiaCsoAvailabilities(qTypes, moniPoint, parName, dateFirst, dateLast)
}
\arguments{
\item{qTypes}{vector of data quality type codes, e.g. c("r", "v", "c"): raw,
validated and calibrated data}

\item{moniPoint}{name of monitoring point, e.g. "TEG", "MUE", "STA"}

\item{parName}{name of parameter, e.g. "AFS", "CSB", "CSBf"}

\item{dateFirst}{Date object representing first date to be considered}

\item{dateLast}{Date object representing last date to be considered}
}
\description{
Plots availability of raw and validated data as bar plot into one plot
}
