% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miamdb2.R
\name{hsReadMiaMdbs}
\alias{hsReadMiaMdbs}
\title{Read MIA-CSO databases}
\usage{
hsReadMiaMdbs(root, search.new = FALSE, dbg = FALSE)
}
\arguments{
\item{root}{root directory to start searching for new databases}

\item{search.new}{if TRUE, root directory is searched recursively for new
databases; if FALSE databases are read from R meta database}

\item{dbg}{whether to show debug messages or not}
}
\value{
data frame with columns \code{mdbFile}, \code{mdbDesc}, \code{mdbDir}
}
\description{
Read MIA-CSO databases
}
