% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miamdb2.R
\name{hsUpdateMiaMdbs}
\alias{hsUpdateMiaMdbs}
\title{Update Metadata Database}
\usage{
hsUpdateMiaMdbs(dfMdbs, root)
}
\arguments{
\item{dfMdbs}{data frame with columns \code{mdbDir}, \code{mdbFile} 
containing paths to currently known databases}

\item{root}{path to directory from which to start looking recursively for 
MS Access database files}
}
\value{
Return number n of added database paths
}
\description{
Update Metadata Database
}
