% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelength.R
\name{hsWavelengthAtIndex}
\alias{hsWavelengthAtIndex}
\title{Returns the wavelength in nm that belongs to the given column index i.}
\usage{
hsWavelengthAtIndex(i)
}
\arguments{
\item{i}{Index (i = 1: 200nm, i = 2: 202.5nm, ..., i = 217: 740nm)}
}
\value{
Wavelength corresponding to index
}
\description{
2011-12-19: moved from hsLibFingerprint.r
}
