% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miadir.R
\name{miadir}
\alias{miadir}
\title{Experimental!}
\usage{
miadir(
  owner = NA,
  kind = NA,
  quaLevel = NA,
  resol = NA,
  DS = kwb.misc::hsDirStructure(dbg = dbg),
  depth = 1,
  dbg = FALSE
)
}
\arguments{
\item{owner}{data owner}

\item{kind}{kind of data (hydraulic, quality, rain)}

\item{quaLevel}{quality level (raw, validated, calibrated)}

\item{resol}{resolution}

\item{DS}{directory structure}

\item{depth}{depth}

\item{dbg}{whether to show debug messages or not}
}
\description{
This function does not work! The intension was to get the full path to an MS
Access database file by filtering for certain criteria (owner, kind of data,
quality level (raw, valid, calibrated) and time resolution)...
}
