% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miamdb.R
\name{miamdb}
\alias{miamdb}
\title{Get Path to MS Access Database Used in MIA-CSO}
\usage{
miamdb(kind = NULL, moniPoint = NULL, qua.level = NULL,
  owner = "KWB")
}
\arguments{
\item{kind}{kind of data: "q" = water quality, "h" = hydraulic data, "r" =
rain}

\item{moniPoint}{name of monitoring point, e.g. "STA", "TEG", "MUE"}

\item{qua.level}{data quality level ("r" = raw, "v" = validated, "c" =
calibrated)}

\item{owner}{owner of the data, one of "KWB", "SEN", "BWB"}
}
\value{
This function returns the full path to the Access database containing
  the specified kind of data
}
\description{
Get Path to MS Access Database Used in MIA-CSO
}
