% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miamdb2.R
\name{miamdb2}
\alias{miamdb2}
\title{Full Path to MIA CSO Database File}
\usage{
miamdb2(id = 0)
}
\arguments{
\item{id}{optional. Integer number identifying the database file. If not 
given the user is asked to enter a number on the console.}
}
\value{
full path to database file or empty string \code{""} if an invalid
  \code{id} was given.
}
\description{
Full Path to MIA CSO Database File
}
