% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miamdb.R
\name{qmdb}
\alias{qmdb}
\title{MS Access Databases (Water Quality Data)}
\usage{
qmdb(...)
}
\arguments{
\item{\dots}{arguments passed to \code{\link{miamdb}}}
}
\description{
MS Access Databases (Water Quality Data)
}
